API refactory in pm_shared.cpp
- Cached CSPlayer pointer in a global variable
- Changed PM_JumpHeight from lambda function to an inline function
- Changed m_flDuckSpeedMultiplier data type to double

+++ regamedll/public/regamedll/API/CSPlayer.h
@@ -143,7 +143,7 @@ public:
 	double m_flJumpHeight; 
 	double m_flLongJumpHeight; 
 	double m_flLongJumpForce;
-	float m_flDuckSpeedMultiplier;
+	double m_flDuckSpeedMultiplier;
 };
 
 // Inlines
